Started the Greek translation
This is my first time trying to translate anything, and even though Greek is my native tongue I can't be 100% confident that I've chosen the best possible translations. Everything is accurate as far as I can tell, but a more experienced translator might be able to point out some flaws and/or improve upon my work.

Notes: "Diary" is translated as "Προσωπικό ημερολόγιο" while "Calendar" is translated as "Ημερολόγιο". Sadly there isn't another word for "diary"to my knowledge.
I've added "(browser)" in parentheses in line 63 as browser is actually the commonly used word for browser instead of the Greek counterpart.